fix(tests): merge forwarded badge tests to fix CI fetch failure
Vite's browser dev server intermittently fails to serve a test file
when many files are requested concurrently in CI. Moving the forwarded
header badge tests into the existing badges.test.jsx (which already
imports MessageItem and is proven stable in CI) eliminates the extra
file that was failing to load.
